+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * :fire: Base64x encodeToBytes method renamed to encode, decodeToBytes method renamed to decode
 * :fire: Base64x removes encodeTo, encodeToBuffer, decodeTo, decodeToBuffer and all methods whose method names begin with 'url' or 'mime'
 * :fire: Remove Mapx.EMPTY_MAP, Collectionx.EMPTY_LIST, Collectionx.EMPTY_SET
+* :bug: Collectionx avoids using java.util.ArrayList because it can't be added
 
 ## 0.6.6